+++ src/utils/agent/utils.ts
@@ -1,3 +1,4 @@
+import { AIMessage, HumanMessage, ToolMessage } from '@langchain/core/messages';
 import { RawMessage } from './index'
 
 // Types for agent message processing
@@ -35,8 +36,9 @@ export function extractMessageContent(content: any): string {
  * This function finds the last user message and extracts only relevant messages after it,
  * then categorizes them into different types (message, thought, tool_call, tool_result)
  */
-export function parseAgentMessages(rawMessages: RawMessage[]): AgentMessage[] {
+export function parseAgentMessages(rawMessages: (AIMessage | HumanMessage | ToolMessage)[]): AgentMessage[] {
   if (!Array.isArray(rawMessages) || rawMessages.length === 0) {
+    console.log('not array')
     return [];
   }
 
@@ -52,65 +54,67 @@ export function parseAgentMessages(rawMessages: RawMessage[]): AgentMessage[] {
 
   // Extract only messages that came after the last user message
   const relevantMessages = rawMessages.slice(userMessageIndex + 1);
+
+  for (const msg of relevantMessages) {
+    console.log('relevant message', typeof msg, msg)
+  }
   
   // Categorize messages
   return relevantMessages
     .map((msg: any) => {
       // Skip any non-constructor messages or human messages
-      if (msg.type !== 'constructor' || msg.id.includes('HumanMessage')) {
+      if (msg instanceof HumanMessage) {
+
         return null;
       }
       
       // Handle AI messages
-      if (msg.id.includes('AIMessage')) {
+      if (msg instanceof AIMessage) {
+        console.log('AI message', msg)
         // Get the content, handling both string and array formats
-        const content = extractMessageContent(msg.kwargs.content);
+        const content = extractMessageContent(msg.content);
         
         // Skip temporary processing messages
         if (content === AGENT_MESSAGES.ERROR || content === "Processing your message...") {
           return null;
         }
         
-        // Get all AI messages that are not processing messages
-        const nonProcessingAIMessages = relevantMessages.filter((m: any) => 
-          m.id.includes('AIMessage') && 
-          extractMessageContent(m.kwargs.content) !== "Processing your message..." &&
-          extractMessageContent(m.kwargs.content) !== AGENT_MESSAGES.ERROR
-        );
-        
         // The last message is the final response
-        const isLastMessage = nonProcessingAIMessages.indexOf(msg) === nonProcessingAIMessages.length - 1;
-        const hasToolCalls = msg.kwargs?.tool_calls?.length > 0;
+        const isLastMessage = relevantMessages.indexOf(msg) === relevantMessages.length - 1;
+        const hasToolCalls = msg.tool_calls?.length ?? 0 > 0;
         
         // If it has tool calls, it's a tool call message
         if (hasToolCalls) {
           return {
             type: 'tool_call' as const,
-            content: extractMessageContent(msg.kwargs.content),
+            content: extractMessageContent(msg.content),
             metadata: {
-              tool_calls: msg.kwargs.tool_calls,
-              usage_metadata: msg.kwargs.usage_metadata
+              tool_calls: msg.tool_calls,
+              usage_metadata: msg.usage_metadata
             }
           };
         }
         
         // If it's the last message, it's the final response
         if (isLastMessage) {
+          console.log('last message', msg)
           return {
             type: 'message' as const,
-            content: extractMessageContent(msg.kwargs.content),
+            content: extractMessageContent(msg.content),
             metadata: {
-              usage_metadata: msg.kwargs.usage_metadata
+              usage_metadata: msg.usage_metadata
             }
           };
         }
         
         // Otherwise it's a thought
+        console.log('thought', msg)
+
         return {
           type: 'thought' as const,
-          content: extractMessageContent(msg.kwargs.content),
+          content: extractMessageContent(msg.content),
           metadata: {
-            usage_metadata: msg.kwargs.usage_metadata
+            usage_metadata: msg.usage_metadata
           }
         };
       }
@@ -135,7 +139,10 @@ export function parseAgentMessages(rawMessages: RawMessage[]): AgentMessage[] {
 /**
  * Process agent response to extract final response and structured messages
  */
-export function processAgentResponse(messages: RawMessage[]): ParsedAgentResponse {
+export function processAgentResponse(messages: (HumanMessage | AIMessage | ToolMessage)[]): ParsedAgentResponse {
+
+  console.log('messages is array', Array.isArray(messages), messages.length)
+
   if (!messages) {
     return {
       response: AGENT_MESSAGES.ERROR,
@@ -144,6 +151,7 @@ export function processAgentResponse(messages: RawMessage[]): ParsedAgentRespons
   }
   
   const processedMessages = parseAgentMessages(messages);
+  console.log('processedMessages in process agent response', processedMessages.length)
   
   // The final response is the last message of type 'message'
   const finalResponse = processedMessages
